Feat: locol 파이썬에 출장 정보 넘기는 기능

+++ src/main/java/team/mjk/agent/domain/prompt/dto/response/Hotel.java
@@ -7,7 +7,7 @@ public record Hotel(
     String arrival_date,
     String destination,
     int guests,
-    long budget,
+    int budget,
     List<String> requirements
 ) {
